Proxy ping and pong messages
This is a fork of https://github.com/koding/websocketproxy, which we're
using inside our API gateway to proxy websocket connections. The upstream
package doesn't proxy ping and pong messages, so our connections are always
timing out after 1 minute. The default ping and pong handlers for the
gorilla/websocket package do exactly what you'd expect - respond with a pong
when a ping is recevied and do nothing when a pong is received. This makes sense
for clients and servers, but a proxy should route these messages to the other
end of the proxied connection so we can keep the connection alive.